requirements.txt
flask>=3.0.0
peewee>=3.17.0
boto3>=1.34.0
requests>=2.31.0
python-dotenv>=1.0.0
apscheduler>=3.10.0
anthropic>=0.40.0
